ingester_server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
